remove BiologicalSex from SprenUI Config into separate util
